# kanban-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban Board API",
        "description": "API para gestionar Tableros, Tarjetas y Tareas de un sistema Kanban, siguiendo la especificación OpenAPI 3.",
        "contact": {
            "name": "Tu Nombre",
            "email": "tu.email@ejemplo.com"
        },
        "license": {
            "name": "Licencia Estándar de API",
            "url": "http://localhost/licencia"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/api/v1/boards/{boardId}/cards": {
            "get": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Obtiene todas las tarjetas (columnas) de un tablero específico",
                "operationId": "dbaa4d87bf14dafb5875b6db094aabeb",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tarjetas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Crea una nueva tarjeta (columna) para el tablero especificado",
                "operationId": "4b10eb5924d4b5f4573351c778d9a73a",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "ID del tablero al que se añadirá la tarjeta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la tarjeta a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "order"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "En Progreso"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarjeta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/cards/{id}": {
            "get": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Obtiene una tarjeta específica por ID",
                "operationId": "7b398c0e2cb012db06205baa0579bd51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarjeta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarjeta obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarjeta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Actualiza el título o la posición de una tarjeta (columna)",
                "description": "Se usa para renombrar la columna o para reordenar las columnas dentro del mismo tablero.",
                "operationId": "85ff55289e5a41500a52a1d683e23409",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarjeta a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos para actualizar",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Listo para QA"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarjeta actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarjeta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Elimina una tarjeta (columna) por su ID",
                "operationId": "ec04f9605058fb5bfc97859fdd9ea7ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarjeta a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarjeta eliminada exitosamente (No Content)"
                    },
                    "404": {
                        "description": "Tarjeta no encontrada"
                    }
                }
            }
        },
        "/api/v1/cards/{id}/move": {
            "patch": {
                "tags": [
                    "Tarjetas"
                ],
                "summary": "Mueve una tarjeta a otro tablero",
                "description": "Método específico para actualizar el board_id de una tarjeta, lo que la mueve a otro tablero.",
                "operationId": "38710aa41711e7cdabe9cfc723199a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarjeta a mover",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nuevo ID del tablero de destino",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "board_id"
                                ],
                                "properties": {
                                    "board_id": {
                                        "description": "El ID del nuevo tablero al que pertenece la tarjeta.",
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ID de tablero actualizado exitosamente"
                    },
                    "404": {
                        "description": "Tarjeta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Card": {
                "title": "Card",
                "description": "Modelo de datos de una Tarjeta (Card), que actúa como columna dentro de un Tablero (Board).",
                "properties": {
                    "id": {
                        "description": "ID único de la tarjeta/columna",
                        "type": "integer"
                    },
                    "board_id": {
                        "description": "ID del tablero al que pertenece la tarjeta",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Título de la tarjeta (ej. Por Hacer, En Progreso)",
                        "type": "string"
                    },
                    "order": {
                        "description": "Posición de la tarjeta dentro del tablero (para ordenar las columnas)",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}